move cli to root
