+++ .gitignore
@@ -33,3 +33,7 @@ go.work.sum
 
 
 bin
+
+downloads
+
+extracted
