--- cli/go.mod
@@ -1,5 +1,0 @@
-module github.com/UmbrellaCrow612/binman/cli
-
-go 1.24.6
-
-require github.com/goccy/go-yaml v1.19.2
