+++ global/console.go
