+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/UmbrellaCrow612/binman
+
+go 1.24.6
